requirements.txt
SPARQLWrapper
gridappsd-python>=1.0
